data_transform/pyproject.toml
[project]
name = "data-transform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-storage-blob>=12.25.0",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "pydantic-settings>=2.8.1",
]
